[FIX] 북마크 삭제 시 해당 북마크가 존재하는지 확인

+++ src/main/java/com/spoony/spoony_server/adapter/out/persistence/zzim/ZzimPersistenceAdapter.java
@@ -107,8 +107,11 @@ public class ZzimPersistenceAdapter implements ZzimPostPort {
     public void deleteByUserAndPost(User user, Post post) {
         PostEntity postEntity = postRepository.findById(post.getPostId())
                 .orElseThrow(() -> new BusinessException(PostErrorMessage.POST_NOT_FOUND));
-        postEntity.updateZzimCount(-1);
-        postRepository.save(postEntity);
-        zzimPostRepository.deleteByUser_UserIdAndPost_PostId(user.getUserId(), post.getPostId());
+        boolean exists = zzimPostRepository.existsByUser_UserIdAndPost_PostId(user.getUserId(), post.getPostId());
+        if (exists) {
+            postEntity.updateZzimCount(-1);
+            postRepository.save(postEntity);
+            zzimPostRepository.deleteByUser_UserIdAndPost_PostId(user.getUserId(), post.getPostId());
+        }
     }
 }
